backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
langchain-core==0.3.79
langchain-nvidia-ai-endpoints==0.3.18
pypdf==6.1.1
python-docx==1.2.0
